refactor: handle exceptions with prefix
