+++ src/main/java/lotto/util/BaseProblem.java
@@ -1,25 +1,23 @@
 package lotto.util;
 
-enum BaseProblem {
+public interface BaseProblem {
+    String message();
+
+    RuntimeException exception();
+}
+
+enum BaseProblemImpl implements BaseProblem {
     PARSE_LONG_FAILED,
     PARSE_INTEGER_FAILED,
     ;
 
-    // TODO 이거 전부 ExceptionHandler로 처리하기
-    private static final String ERROR_PREFIX = "[ERROR]";
-
-    IllegalArgumentException exception() {
-        return new IllegalArgumentException(this.toString());
-        // return Global.EXCEPTION_HANDLER.of(this).log().raiseException();
+    @Override
+    public String message() {
+        return this.name();
     }
 
     @Override
-    public String toString() {
-        return String.format("%s %s", ERROR_PREFIX, this.name());
+    public RuntimeException exception() {
+        return Global.EXCEPTION_HANDLER.exception(this);
     }
-
-    IllegalArgumentException exception(Throwable e) {
-        return new IllegalArgumentException(this.toString(), e);
-    }
-
 }
